AN-remove state management of the users selected skills from the SelectedSkills component and create a context for the selected skills which can be used to pass state to multiple components
